+++ gizmo_toolpreset_bar.py
@@ -71,7 +71,6 @@ class STORYTOOLS_GGT_toolpreset_bar(GizmoGroup):
             op.layer = props.layer
             op.material = props.material
             op.brush = props.brush
-            op.brush_asset_path = props.brush_asset_path
             op.description = props.description
             op.shortcut = kmi.to_string() # Shortcut text for description
             self.tool_preset_gizmos.append(gz)